Refactoring almost complete, doesn't compile yet because some structures aren't created yet
