+++ structures/index.h
@@ -1,2 +1,7 @@
+#ifndef STRUCTURES_INDEX_H
+#define STRUCTURES_INDEX_H
+
 #include "./user.h"
 #include "./question.h"
+
+#endif
