+++ utils/index.h
@@ -1,4 +1,11 @@
+#ifndef UTILS_INDEX_H
+#define UTILS_INDEX_H
+
 #include "./io.h"
 #include "./string.h"
 #include "./question.h"
 #include "./file.h"
+#include "./quiz.h"
+
+
+#endif
